get new pytests to work
